pyproject.toml
[tool.poetry]
name = "1hourdev"
version = "0.1.0"
description = ""
authors = ["Aankit Roy"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
annotated-types = "0.5.0"
anyio = "3.7.1"
bcrypt = "4.1.2"
cffi = "1.15.1"
click = "8.1.7"
cryptography = "42.0.2"
dnspython = "2.3.0"
ecdsa = "0.18.0"
exceptiongroup = "1.2.0"
fastapi = "0.103.2"
h11 = "0.14.0"
idna = "3.6"
importlib-metadata = "6.7.0"
motor = "3.3.2"
passlib = "1.7.4"
pyasn1 = "0.5.1"
pycparser = "2.21"
pydantic = "2.7.1"
pydantic-core = "2.18.2"
pyjwt = "2.8.0"
pymongo = "4.6.1"
python-dotenv = "0.21.1"
python-jose = "3.3.0"
python-multipart = "0.0.6"
rsa = "4.9"
six = "1.16.0"
sniffio = "1.3.0"
starlette = "0.27.0"
typing-extensions = "4.7.1"
uvicorn = "0.22.0"
zipp = "3.15.0"
openai = "1.16.1"
fastapi-mail = "^1.4.1"
pydantic-settings = "^2.2.1"
boto3 = "^1.34.100"
reportlab = "^4.2.0"
requests = "^2.32.3"
httpx = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
